requirements.txt
aws-cdk.core>=1.132.0
aws-cdk.assertions>=1.132.0
aws-cdk.aws_kinesisfirehose>=1.132.0
aws-cdk.aws_kinesisfirehose_destinations>=1.132.0
aws-cdk.aws_glue>=1.132.0
aws-cdk.aws_s3>=1.132.0
aws-cdk.aws_s3_notifications>=1.132.0
aws-cdk.aws_iam>=1.132.0
aws-cdk.aws_events>=1.132.0
aws-cdk.aws_events_targets>=1.132.0